Start splitting metadata types - just rename types and dependencies for now
