+++ src/app/data-model-provider/BUILD.gn
@@ -14,6 +14,18 @@
 import("//build_overrides/chip.gni")
 import("//build_overrides/pigweed.gni")
 
+source_set("metadata") {
+  sources = [
+    "MetadataTypes.h",
+  ]
+
+  public_deps = [
+    "${chip_root}/src/access:types",
+    "${chip_root}/src/lib/core:types",
+    "${chip_root}/src/lib/support",
+  ]
+}
+
 source_set("data-model-provider") {
   sources = [
     "ActionContext.h",
@@ -32,7 +44,7 @@ source_set("data-model-provider") {
   ]
 
   public_deps = [
-    "${chip_root}/src/access:types",
+    ":metadata",
     "${chip_root}/src/app:attribute-access",
     "${chip_root}/src/app:command-handler-interface",
     "${chip_root}/src/app:events",
